+++ packages/interfaces-adt/src/auth/AuthType.ts
@@ -1,4 +1,0 @@
-/**
- * Authentication type
- */
-export type AuthType = 'jwt' | 'xsuaa' | 'basic';